refactoring for makefile
