+++ .env
@@ -1,17 +1,17 @@
 #############################################
 # PostgreSQL database environment variables
 #############################################
-DATABASE_HOST=database_sample
-DATABASE_ADMIN=postgres
-DATABASE_ADMIN_PASSWORD=pgpswd
-DATABASE_DEFAULT_NAME=postgres
+; DATABASE_HOST=database_sample
+; DATABASE_ADMIN=postgres
+; DATABASE_ADMIN_PASSWORD=pgpswd
+; DATABASE_DEFAULT_NAME=postgres
 
-DATABASE_USER="admin_tracker"
-DATABASE_PASSWORD=pgpswd
-DATABASE_NAME=tracker
-DATABASE_ROLE=tracker
-DATABASE_PORT=5432
+; DATABASE_USER="admin_tracker"
+; DATABASE_PASSWORD=pgpswd
+; DATABASE_NAME=tracker
+; DATABASE_ROLE=tracker
+; DATABASE_PORT=5432
 
-SAMPLE_SERVICE_HOST=0.0.0.0
-SAMPLE_SERVICE_PORT=8000
-SAMPLE_SERVICE_PATH="sample"
+; SAMPLE_SERVICE_HOST=0.0.0.0
+; SAMPLE_SERVICE_PORT=8000
+; SAMPLE_SERVICE_PATH="sample"
