+++ sampleService/docker-compose.testing.yml
@@ -42,7 +42,7 @@ services:
     restart: always
     command: "sh -c 'uvicorn main:app --reload --host 0.0.0.0 --port 8000'"
     volumes:
-      - ./src:/app
+      - ./src:/app/src
     ports:
       #- "${SAMPLE_SERVICE_PORT}:${SAMPLE_SERVICE_PORT}"
       - 8000:8000
