+++ sampleService/src/Dockerfile
@@ -1,7 +1,7 @@
 FROM tiangolo/uvicorn-gunicorn-fastapi:python3.11-slim
 
 ENV PYTHONUNBUFFERED=1
-WORKDIR /app
+WORKDIR /app/src
 COPY requirements.txt .
 
 RUN pip install --no-cache-dir -r requirements.txt
