+++ sampleService/src/db/session.py
@@ -1,6 +1,5 @@
-''' Db session. '''
-from core.config import settings
-#from sqlalchemy.ext.declarative import declarative_base
+""" Creating db session. """
+from src.core.config import settings
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
 
@@ -21,6 +20,3 @@ def get_db():
         yield db
     finally:
         db.close()
-
-# def init_db():
-#     Base.metadata.create_all(bind=engine)
