+++ sampleService/src/main.py
@@ -2,8 +2,8 @@ from fastapi import FastAPI, HTTPException, Request
 from fastapi.responses import JSONResponse
 from pydantic import ValidationError
 from sqlalchemy.exc import NoResultFound
-from resources.errors import UserNotFound
-from api import sample
+from src.resources.errors import UserNotFound
+from src.api import sample
 
 app = FastAPI()
 
@@ -11,17 +11,16 @@ app.include_router(sample.router, prefix="/sample", tags=["sample"])
 
 @app.exception_handler(HTTPException)
 def validation_exception_handler(exc: ValidationError):
-    
+
     # I can add input parameter "request: Request" for flexible logging.
     return JSONResponse(
         status_code=422,
         content={"detail": exc.json()},
     )
-    
+
 @app.exception_handler(HTTPException)
 def notfound_exception_handler(request: Request, exc: NoResultFound):
-    
-    # I can add input parameter "request: Request" for flexible logging.
+
     return JSONResponse(
         status_code=404,
-        content={"detail": UserNotFound.DETAIL},)  
+        content={"detail": UserNotFound.DETAIL},)
