--- sampleService/tests/component/pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-testpaths= /Users/bo09886/Repos/Tracker/sampleService/tests/component
