+++ sampleService/tests/unit/conftest.py
@@ -7,15 +7,13 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from typing_extensions import Annotated
 
-from sampleService.tests import config
+from tests import config
 
 def get_settings():
     return config.Settings()
 
 def get_db_url(settings: Annotated[config.Settings, Depends(get_settings)]) -> PostgresDsn:
     ''' Generate db url. '''
-    print(settings.database_user)
-    print('*********')
     try:
         conn = PostgresDsn.build(
             scheme='postgresql',
