+++ sampleService/tests/unit/helpers/api_helper.py
@@ -1,8 +1,8 @@
 """ Helpers for working with API. """
 
-from sampleService.src.api import sample
-from sampleService.src.schemas import schemas
-from sampleService.tests.unit.helpers.structures import CreateUserResult
+from src.api import sample
+from src.schemas import schemas
+from tests.unit.helpers.structures import CreateUserResult
 
 def post_user(db, username) -> CreateUserResult:
     result = sample.test_post(schemas.UserCreate(username=username), db)
